+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.3.0](https://github.com/dldevinc/paper-uploads/tree/v0.3.0) - 2020-07-22
 - Refactored class hierarchy
+- Rename `MimetypeValidator` to `MimeTypeValidator`
 - Signal `variation_created` was added
 - Use DropBox checksum realization.
 - Remove postprocessing